Rename `unicode` import to `es2015`
