+++ .gitignore
@@ -1,6 +1,6 @@
 /index.js
 /text.js
-/unicode
+/es2015
 
 # Coverage report
 coverage
