+++ src/es2015/index.js
