+++ src/es2015/text.js
